% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_bigeye_data.R
\name{process_bigeye_data}
\alias{process_bigeye_data}
\title{Function to process bigeye data
Process three types of bigeye data: annual counts, annual counts by length bins, 
annual counts by length bins and seasons}
\usage{
process_bigeye_data(nspp)
}
\arguments{
\item{nspp}{Number of species to divide length bins into}
}
\description{
Function to process bigeye data
Process three types of bigeye data: annual counts, annual counts by length bins, 
annual counts by length bins and seasons
}
