% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vast.R
\name{run_vast}
\alias{run_vast}
\title{Function run VAST
Specify model characteristics outside of the function. Currently has the ability
to save the output and write the results to the directory}
\usage{
run_vast(dat_index, ObsModel)
}
\arguments{
\item{dat_index}{Specify the dataset to use numerically. Options are 1-annual counts; 
2-annual numbers by size; 3 - annual numbers by size and season}

\item{ObsModel}{ObsModel configuration}
}
\description{
Function run VAST
Specify model characteristics outside of the function. Currently has the ability
to save the output and write the results to the directory
}
